# Add `broadcast!(identity, ::Array, ::Array)` specialization