Closes #20227.

Performance improvement is most prominent for small element size:
```julia
x=rand(Int8,50,50); y=similar(x); @benchmark $y .= $x

# master
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     9
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     2.20 μs (0.00% GC)
  median time:      2.21 μs (0.00% GC)
  mean time:        2.22 μs (0.00% GC)
  maximum time:     3.38 μs (0.00% GC)

# this PR
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     976
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     70.00 ns (0.00% GC)
  median time:      70.00 ns (0.00% GC)
  mean time:        70.19 ns (0.00% GC)
  maximum time:     130.00 ns (0.00% GC)
```

There is a little bit of overhead though, becoming significant for very small arrays:
```julia
x=rand(Int64,1); y=similar(x); @benchmark $y .= $x

# master
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     999
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     9.00 ns (0.00% GC)
  median time:      10.00 ns (0.00% GC)
  mean time:        9.96 ns (0.00% GC)
  maximum time:     19.00 ns (0.00% GC)

# this PR
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     15.00 ns (0.00% GC)
  median time:      15.00 ns (0.00% GC)
  mean time:        15.21 ns (0.00% GC)
  maximum time:     34.00 ns (0.00% GC)
```