@martinholters 2017-01-26
Replaced the new specialization with the shortened version proposed by @Sacha0.